pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "apigen"
authors = [{ name = "Artyom Fomenko", email = "artyom.fomenko@distribusion.com" }]
readme = "README.md"
dynamic = ["version", "description"]
dependencies = [
    "pydantic",
    "pyyaml",
    "typer",
    "importlib_resources",
    "jinja2"
]

[project.scripts]
apigen = "apigen.main:main"


[tool.isort]
profile = "black"


[tool.mypy]
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
implicit_reexport = false
strict_equality = true


[tool.pylint.messages_control]
disable = "C0115, R0903, C0116, R1734, R1735"


[tool.pylint.format]
max-line-length = "120"


[tool.black]
include = "apigen"


[tool.coverage.run]
source = ['apigen']


[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
asyncio_mode = "strict"
